Move make_blind_index into the blind index adapter package
Place the factory in pydantic_encryption/adapters/blind_index/factory.py
alongside the blind index backends and export it from that package, keeping
the package root for cross-cutting modules only.
