+++ pydantic_encryption/adapters/blind_index/__init__.py
@@ -1,7 +1,9 @@
 from pydantic_encryption.adapters.blind_index.argon2 import Argon2BlindIndexAdapter
+from pydantic_encryption.adapters.blind_index.factory import make_blind_index
 from pydantic_encryption.adapters.blind_index.hmac_sha256 import HMACSHA256Adapter
 
 __all__ = [
     "HMACSHA256Adapter",
     "Argon2BlindIndexAdapter",
+    "make_blind_index",
 ]
